requirements.txt
certifi~=2021.5.30
ExifRead~=2.3.2
Pillow~=8.3.1
